Moved some routes from gamerRoutes to postRoutes
