+++ frontend/src/Services/PostServices.ts
@@ -3,6 +3,10 @@ import axios from 'axios';
 const POST_BASE_URL = 'http://localhost:5000/posts';
 
 class PostServices {
+  createPost(data: { name: string; content: string }) {
+    return axios.post(POST_BASE_URL, data);
+  }
+
   getPosts() {
     return axios.get(POST_BASE_URL);
   }
